Draft `java_toolchain_resolution`
This Bazel workspace includes:

- A JDK13 external repository (i.e. `http_archive`).
- A `java_runtime_toolchain` rule to wrap this JDK.
- A `java_toolchain` rule to wrap this JDK with some "do nothing" tools.
(The use of the "do nothing" `cc_binary` was adapted from the
`dummy_java_toolchain` in `@bazel_tools`.)
- A `register_toolchains` rule which takes both of the above toolchains.
- The `--incompatible_use_toolchain_resolution_for_java_rules` flag set
in the `.bazelrc` to try to enable toolchain resolution.
- A trivial `java_binary` which just prints out the `java.version` at
run-time.

Unfortunately, with Bazel v1.0.0rc3, these two toolchains don't seem to
be used to build and run this `java_binary`, even when the
`--toolchain_resolution_debug` flag indicates that these custom
toolchains are being selected. Maybe implementation of this feature
isn't finished yet.

I should try this again with a later Bazel release.